requirements.txt
flask==2.2.3
tensorflow-cpu==2.10.0
pillow==9.2.0
